rename: main -> bookmark path 변경
